feat: implement webview for user interface and router for passing data into webview
