+++ app.json
@@ -11,6 +11,7 @@
       "resizeMode": "contain",
       "backgroundColor": "#ffffff"
     },
+    "scheme": "collectiv",
     "assetBundlePatterns": [
       "**/*"
     ],
